ROU-3474: Creating RowStyleInfo type
